Add: Add output interface, Update monitor NDA transition logic, No test Add open-loop test method
Why: Add output interface

How: Add output interface

+++ build/build.sh
@@ -1,7 +1,23 @@
 #!/bin/bash -e
 
-find . ! \( -name '*.sh' -or -name '..'  -or -name '.' -or -name '.gitignore' \) | xargs rm -rf
+function help() 
+{
+    echo "sh build.sh: normal build"
+    echo "sh build.sh [open_test]: build with open_test"
+}
 
-cmake .. && make -j8
-# make clean 
-# make -j8
+function main() 
+{
+    if [[ $1 == '--help' || $1 == '-h' ]]; then
+        help    
+        exit 0
+    fi
+    find . ! \( -name '*.sh' -or -name '..'  -or -name '.' -or -name '.gitignore' \) | xargs rm -rf
+    if [[ $# -eq 1 && $1 == 'open_test' ]]; then
+        cmake .. -DOPEN_TEST=ON && make -j8
+        exit 0
+    fi
+    cmake .. && make -j8
+}
+
+main $@
